egl/dri: Add gbm platform
This adds a GBM platform that doesn't depend on gbm's dri backend. This
can be used, for instance, with the intel backend.

This also adds a gbm_bo_export entrypoint to gbm, so that the EGL code
can access the buffer. This is necessary since EGL won't share the fd
width gbm when using the new platform.